move episode list into episodeswitcher
